pyproject.toml
[tool.poetry]
name = "fastapimongodb"
version = "0.1.0"
description = ""
authors = ["Roman Volodin <rozarioagro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
beanie = "^1.25.0"
pydantic-settings = "^2.2.1"
python-slugify = "^8.0.4"
orjson = "^3.9.15"
fastapi-pagination = "^0.12.19"
shortuuid = "^1.0.13"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
flake8-bugbear = "^24.1.17"
flake8-isort = "^6.1.1"
flake8-bandit = "^4.1.1"
flake8-pytest-style = "^1.7.2"
flake8-pytest = "^1.4"
flake8-print = "^5.0.0"
flake8-commas = "^2.1.0"
flake8-broken-line = "^1.0.0"
isort = "^5.13.2"

[tool.poetry.group.test.dependencies]
pytest = "^8.0.0"
faker = "^22.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
